+++ src/text/index.unit.tsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import renderer from 'react-test-renderer'
 import Text from 'text'
 
 const multipleLineText = `line1
@@ -6,10 +7,19 @@ line2
 line3`
 
 describe('Text', () => {
-  it('Should have the default classes', () => {
-    const wrapper = shallow(<Text>{'test'}</Text>)
-    expect(wrapper.hasClass('kirk-text')).toBe(true)
-    expect(wrapper.hasClass('kirk-text-body')).toBe(true)
+  it('should have change the DOM', () => {
+    const text = renderer.create(<Text>text</Text>).toJSON()
+    expect(text).toMatchSnapshot()
+  })
+  it('should have a role and classname', () => {
+    const text = renderer
+      .create(
+        <Text className="custom" role="role">
+          text
+        </Text>,
+      )
+      .toJSON()
+    expect(text).toMatchSnapshot()
   })
 
   it('Should have the custom class', () => {
@@ -25,6 +35,15 @@ describe('Text', () => {
   it('Should render with the div tag', () => {
     const wrapper = shallow(<Text tag="div">{'test'}</Text>)
     expect(wrapper.name()).toBe('div')
+  })
+
+  it('Should have role attribute when present and undefined otherwise', () => {
+    const wrapper = shallow(<Text>{'test'}</Text>)
+    expect(wrapper.prop('role')).toBeUndefined()
+
+    wrapper.setProps({ role: 'alert' })
+
+    expect(wrapper.prop('role')).toBe('alert')
   })
 
   it('Should accept ReactNode as children', () => {