+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - **[NEW]** `SuccessModal` component
+- **[UPDATE]** Move keycodes into a single util
 - **[UPDATE]** Add possibility to have a role attribute on `<Text>`
 - **[UPDATE]** Add the role "alert" on `<Snackbar>` for accessibility
 - **[UPDATE]** Add aria-hidden on the `<Textfield>` clear button